+++ Source/SaveResult.swift
@@ -22,7 +22,7 @@ import Foundation
 /**
  A result object representing the result of saving an `NSManagedObjectContext`.
  */
-public enum SaveResult {
+public enum SaveResult: Equatable {
 
     /// The success result.
     case success
@@ -40,21 +40,5 @@ public enum SaveResult {
             return error
         }
         return nil
-    }
-}
-
-extension SaveResult: Equatable {
-    /// :nodoc:
-    public static func == (lhs: SaveResult, rhs: SaveResult) -> Bool {
-        switch (lhs, rhs) {
-        case (.success, .success):
-            return true
-
-        case (let .failure(error1), let .failure(error2)):
-            return error1 == error2
-
-        default:
-            return false
-        }
     }
 }